Enumerate the loops that wanted both the index and the element
Seventy-one loops walked a container with c10::irange over its size, indexed
only that container, and needed the counter for something else as well --
passing it to a C API, comparing it against a bound. std::views::enumerate
hands over both, so the subscript goes away while the index stays.

The scan that found them looked for the whole shape: the loop variable used
nowhere but as cont[i], plus at least one other use, and no arithmetic in any
subscript. Of the 75 it first matched, one was a loop nested inside another
match and is left alone, and three could not be converted after all: ivalue.cpp
formats a container reached through a template parameter, which is not required
to be a viewable_range, and basic.cpp iterates an at::Tensor, which is not a
range at all.

enumerate yields the index as the range's difference_type, which is signed
where irange's was size_t, so four comparisons had to move with it.
code_template.h and TensorIndexing.cpp now compare against std::ssize; the
latter also stops computing size() - 1 in unsigned arithmetic, which underflows
on an empty container. ivalue_inl.h's compaction cursor becomes ptrdiff_t, and
ProcessGroupGloo drops a cast to size_t from rootTensor, which was signed to
begin with.

Test Plan:

Every changed file, and a translation unit including each changed header --
thirty-seven compilations, from their own commands in compile_commands.json,
run ten at a time:

```
xargs -P 10 -n 1 bash < jobs.txt
```

All clean.

+++ aten/src/ATen/test/basic.cpp
@@ -19,6 +19,8 @@ struct THFloatTensor;
 #include <thread>
 #include <mutex>
 
+#include <ranges>
+
 #define ASSERT_EQ_RESOLVED(X, Y) \
   {                              \
     bool isEQ = X == Y;          \
@@ -256,13 +258,13 @@ void TestIndexingByScalar() {
   Tensor tensor = arange(0, 10, kInt);
   Tensor one = ones({}, kInt);
   for (const auto i : c10::irange(tensor.numel())) {
-    ASSERT_TRUE(tensor[i].equal(one * i));
+    ASSERT_TRUE(tensor[i].equal(one * static_cast<int64_t>(i)));
   }
   for (size_t i = 0; i < static_cast<uint64_t>(tensor.numel()); ++i) {
     ASSERT_TRUE(tensor[i].equal(one * static_cast<int64_t>(i)));
   }
   for (const auto i : c10::irange(tensor.numel())) {
-    ASSERT_TRUE(tensor[i].equal(one * i));
+    ASSERT_TRUE(tensor[i].equal(one * static_cast<int64_t>(i)));
   }
   // NOLINTNEXTLINE(bugprone-too-small-loop-variable)
   for (int16_t i = 0; i < tensor.numel(); ++i) {
